requirements.txt
apify-client>=1.6.0
anthropic>=0.18.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
